+++ classes/controller.sauce.class.php
@@ -233,13 +233,6 @@ class sauce_lib {
 
     private function __clone() {}
 
-    /**
-    * Private unserialize method to prevent unserializing of the *Singleton*
-    * instance.
-    *
-    * @return void
-    */
-
-    private function __wakeup() {}
+    public function __wakeup() {}
 
 }