+++ classes/post_exclude.class.php
@@ -46,13 +46,6 @@ class post_exclude {
 
   private function __clone() {}
 
-  /**
-  * Private unserialize method to prevent unserializing of the *Singleton*
-  * instance.
-  *
-  * @return void
-  */
-
-  private function __wakeup() {}
+  public function __wakeup() {}
 
 }